Extract embedded-path write strategies into their own module
Move the embedded-write machinery (status update, compensation, embedded
balance effects, creation/transition strategies) out of update-status.js
so each file stays under the 500-line limit. update-status.js keeps the
ledger path and the dispatch orchestration. Behaviour is unchanged; also
reattach the JSDoc that had drifted off updatePrnStatus and the in-memory
repository factory.
